+++ front/src/components/header/DarkMode.jsx
@@ -1,6 +1,4 @@
 import React from "react";
-import { Link } from "react-router-dom";
-import imgsun from "../../assets/images/icon/sun.png";
 
 const DarkMode = () => {
   let clickedClass = "clicked";
@@ -31,16 +29,7 @@ const DarkMode = () => {
       theme = darkTheme;
     }
   };
-  return (
-    <div className="mode_switcher">
-      <h6>
-        Dark mode <strong>Available</strong>
-      </h6>
-      <Link to="#" onClick={(e) => switchTheme(e)}>
-        <img src={imgsun} alt="" />
-      </Link>
-    </div>
-  );
+  return <></>;
 };
 
 export default DarkMode;
